+++ CHANGELOG.md
@@ -1,13 +1,18 @@
 # Changelog
 
-# 6.327438247
+## Upcoming
+
+- Rewrote in C for funsies.
+- Now restores terminal state on shutdown.
+
+## 6.327438247
 
 - Relicensed as GPLv3+.
 - Removed from PyPI.
 - Removed packaging stuff; BitMasher is now a single script.
 - Code cleanup.
 
-# 5.74351224532
+## 5.74351224532
 
 - "Improved" lose sequence.